Revert "Merge branch 'Tech_Debt_2' of https://github.com/SWE443-TeamRed/open-bank into Tech_Debt_2"
This reverts commit b731c3ee27ff605c3559f8beb86828c845af18ad, reversing
changes made to 032134bccd4e3518aa3f6f5eb211049beb909b64.

+++ open-bank/src/test/java/Test_S12.java
@@ -7,7 +7,6 @@ import org.sdmlib.openbank.Account;
 import org.sdmlib.openbank.User;
 import org.sdmlib.storyboards.Storyboard;
 
-import java.math.BigInteger;
 import java.util.Date;
 public class Test_S12 {
     Account tinaAccount;
@@ -30,7 +29,7 @@ public class Test_S12 {
                 .withName("Tina")
                 .withPassword("1234")
                 .withUserID("Tina");
-        tinaAccount.Account(BigInteger.valueOf(50));
+        tinaAccount.Account(50.00);
         tinaAccount.withOwner(tina)
                 .withCreationdate(new Date());
         //Tina logs in
@@ -51,13 +50,13 @@ public class Test_S12 {
     public void testDeposit() {
         creatingTinaAccount();
         storyboard = new Storyboard();
-        tinaAccount.setBalance(BigInteger.valueOf(50));
+        tinaAccount.setBalance(50.00);
         assert (tina.isLoggedIn());
 
         storyboard.add("Precondition: Tina has $50 in her account and she deposit $30 to her account.");
         storyboard.addObjectDiagram("Tina", tina, "Savings", tinaAccount);
 
-        tinaAccount.deposit(BigInteger.valueOf(30));//Making withdraw
+        tinaAccount.deposit(30.00);//Making withdraw
         storyboard.assertEquals("Tina makes transaction: ",
                 tinaAccount.getBalance(), 50.00 + 30.00);//Making sure it worked.
         storyboard.add("Post-condition: Tina has $80 in her account");
